Return upper pivot
